\name{dqreg}
\alias{dqreg}
\title{Plot}
\description{
Plots posterior density for parametric basis functions.
}
\usage{
dqreg(object, y, X)
}
\arguments{
  \item{object}{
  An object of class qreg.
}
  \item{y}{
  A vector of response values to evaluate the density.
}
  \item{X}{
A vector of predictors.
}
}

